+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.10.3
+
+- Refactor logging for other modules also
+- Fail also if copy_to_local and copy_to_remote is specified for unidirectional sync
+
 ## 0.10.2
 
 - Fine tune exception representation is log messages